Modularize UploadLayout components
